Fixed merge conflict
